+++ src/pages/post/ColorOption.jsx
@@ -3,8 +3,8 @@ import selectedIcon from '../../assets/images/selected.svg';
 
 const ColorOption = ({ color, selected, onClick, className = '' }) => (
   <div
-    className={`relative transition-all cursor-pointer flex items-center justify-center rounded-md ${className}`}
-    style={{ backgroundColor: color.code, width: '100%', height: '100%' }}
+    className={`relative w-full h-full rounded-xl cursor-pointer transition-all overflow-hidden sm:w-32 sm:h-32 md:w-40 md:h-40 lg:w-60 lg:h-60 mb-0 ${className}`}
+    style={{ backgroundColor: color.code }}
     onClick={() => onClick(color.name)}
   >
     {selected && (
